Report fixes
Again, the storage location of the report file has been changed. This time to the user's home directory. And its name to sort_report.txt.

Improved the rules for creating a report, now it is created even if there is nothing to sort. Fixed a bug in create_report which caused that if only one file was sorted it would not save to the report.